Explicitly mark chips as "inReplacedChip".
Current behavior of RecipientChipAddedListener does not distinguish
between chips added from dropdown and from append(). When multiple chips
are added via append(), RecipientEditTextView first creates a
ReplacementChip, and later converts these ReplacementChip to a chip with
contact information. We want to mark these kind of RecipientEntry
explicitly, in order to conclude that RecipientEntry is in a replacement
chip.

Bug: 34804889
Change-Id: I0555e4c6ff1fdaa1a51153034c96d329b778c30a